--- domain/src/main/java/team/aliens/domain/_param/notices/querywhethernewnoticesexist/QueryWhetherNewNoticesExistResponse.kt
@@ -1,10 +1,0 @@
-package team.aliens.domain._param.notices.querywhethernewnoticesexist
-
-/**
- * @author junsuPark
- * A response returned when querying whether new notices exist
- * @property whetherNewNotices a boolean value of whether new notices exists
- */
-data class QueryWhetherNewNoticesExistResponse(
-    val whetherNewNotices: Boolean,
-)
