--- domain/src/main/java/team/aliens/domain/_param/remains/querycurrentappliedremainoption/QueryCurrentAppliedRemainOptionResponse.kt
@@ -1,10 +1,0 @@
-package team.aliens.domain._param.remains.querycurrentappliedremainoption
-
-/**
- * @author junsuPark
- * A response returned when querying current applied remain option
- * @property title the title of the remain option
- */
-data class QueryCurrentAppliedRemainOptionResponse(
-    val title: String,
-)
