+++ domain/src/main/java/team/aliens/domain/_param/school/examineschoolverificationcode/ExamineSchoolVerificationCodeResponse.kt
@@ -1,0 +1,12 @@
+package team.aliens.domain._param.school.examineschoolverificationcode
+
+import java.util.*
+
+/**
+ * @author junsuPark
+ * A response returned when examining school school verification code
+ * @property schoolId a school id can be used at sign up, etc
+ */
+data class ExamineSchoolVerificationCodeResponse(
+    val schoolId: UUID,
+)
