+++ domain/src/main/java/team/aliens/domain/_param/student/editprofile/EditProfileResponse.kt
@@ -1,0 +1,10 @@
+package team.aliens.domain._param.student.editprofile
+
+/**
+ * @author junsuPark
+ * An response returned when editing profile
+ * @property profileImageUrl an url expresses user's profile image url
+ */
+data class EditProfileResponse(
+    val profileImageUrl: String,
+)
